+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - Updated all relic fragment descriptions and file paths for more consistent importing from save files. (Legendary fragments coming soon!)
 - Fixed issue where fragments and fusions could be equipped in scenarios they shouldn't.
 - Added missing fusions and updated text for all fusions.
+- Added a number of new fragments and items to the various calculations in the builder.
 
 ## 2024-09-26